+++ src/z_zone.c
@@ -363,7 +363,7 @@ void Z_ChangeTag2(void *ptr, int tag) {
   if (block->id != ZONEID)
     I_Error("Z_ChangeTag: freed a pointer without ZONEID");
 
-  if (tag >= PU_PURGELEVEL && (unsigned)block->user < 0x100)
+  if (tag >= PU_PURGELEVEL && (unsigned long int)block->user < 0x100)
     I_Error("Z_ChangeTag: an owner is required for purgable blocks");
 
   block->tag = tag;
